edit vacation request view model appended
